Moved slide bar different components
to increase reusability.
